Polish up the mulliken population analysis
What:
1. Modify the code to be a function rather than a class
2. Modify tests for the new code

Why:
1. It will be easier to use (and understand) if the code is written as a
function rather than a class with properties and static methods. This approach
will affect extendibility as modifications of the code will entail hard-coding
specific structure. However, it seems that the main extension may be the use of
arbitrary weights (which was supported in the old code) and this extension is
not necessary at the moment. If necessary, we can simply add an (optional)
parameter for the weights.